requirements-dev.txt
-r requirements.txt

django-extensions==3.2.3
ipython==8.14.0
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2
